finished static and classmethods
