+++ static_and_classmethods/c02.md
@@ -1,3 +1,21 @@
 # Static and Class Methods
 
 ## Exercise 2 (task)
+
+Static methods can sometimes be used for factory generator methods to create instances of another class.
+
+This pattern can be helpful to centralise the creation of classes in order to store them together or apply validation.
+
+Complete the code to show how static methods are used in OOP.
+
+You will notice that is also includes a demonstration of the `@classmethod` decorator. A **classmethod** is also a method bound to the class not the object, but **must** take the *class* as the first argument, written as `cls`. In fact, you could replace the method with the following and it would still work:
+
+```
+  @staticmethod
+  def get_product(name):
+    return Product._registry.get(name)
+```
+
+However using `@classmethod` as shown is better practice because it would allow better for inheritance, e.g. with a class `SpecialProduct(Product)`, which could then override its implementation.
+
+In general, **classmethods** are used for *factory methods* - methods that *return* instances of classes, and **staticmethods** are used for *utility functions* - functions that *belong* in a class for organisational purposes.
