+++ static_and_classmethods/c03.md
@@ -1,3 +1,7 @@
 # Static and Class Methods
 
 ## Exercise 3 (example)
+
+One other common usecase for **classmethods** is writing *alternative constructors*. Since Python *does not* have function overloading, we cannot write more than *one* constructor. Hence, *classmethods* are often used to write alternative constructors that **return** an instance of the class - they are essentially factory methods!
+
+Technically *staticmethods* would also work for this purpose, but it is *not friendly* to inheritance since we would have to **hardcode** the class in the static method. When using a classmethod, we can use use `cls()` to create an instance of the same class.
