--- static_and_classmethods/c04.md
@@ -1,21 +1,0 @@
-# Static and Class Methods
-
-## Exercise 4 (task)
-
-Static methods can sometimes be used for factory generator methods to create instances of another class.
-
-This pattern can be helpful to centralise the creation of classes in order to store them together or apply validation.
-
-Complete the code to show how static methods are used in OOP.
-
-You will notice that is also includes a demonstration of the `@classmethod` decorator.
-
-This has some similarities to using `@staticmethod` and in fact you could replace that method with the following and it would still work:
-
-```
-  @staticmethod
-  def get_product(name):
-    return Product._registry.get(name)
-```
-
-However using `@classmethod` as shown is better practice because it would allow better for inheritance, e.g. with a class `SpecialProduct(Product)`, which could then override its implementation.
